+++ .gitignore
@@ -24,3 +24,4 @@ yarn-error.log*
 
 
 yarn.lock
+package.lock.json
